+++ src/modules/Dashboard/state/Applets/Applets.schema.ts
@@ -31,66 +31,6 @@ export type Applet = {
   updatedAt: string;
 };
 
-export type ActivityFlow = {
-  id: string;
-  name: string;
-  description: string;
-  ordering: number;
-  isSingleReport?: boolean;
-  hideBadge?: boolean;
-  activityIds?: number[];
-  isHidden?: boolean;
-};
-
-export type Activity = {
-  id: string;
-  name: string;
-  description: string;
-  ordering: number;
-  splashScreen?: string;
-  image?: string;
-  showAllAtOnce?: boolean;
-  isSkippable?: boolean;
-  isReviewable?: boolean;
-  responseIsEditable?: boolean;
-  isHidden?: boolean;
-};
-
-type Theme = {
-  id: string;
-  name: string;
-  logo: string;
-  backgroundImage: string;
-  primaryColor: string;
-  secondaryColor: string;
-  tertiaryColor: string;
-  public: boolean;
-};
-
-export type SingleApplet = {
-  id: string;
-  displayName: string;
-  version: string;
-  description: string;
-  about: string;
-  createdAt: string;
-  updatedAt: string;
-  image?: string;
-  watermark?: string;
-  themeId?: string;
-  reportServerIp?: string;
-  reportPublicKey?: string;
-  reportRecipients?: string[];
-  reportIncludeUserId?: boolean;
-  reportIncludeCaseId?: boolean;
-  reportEmailBody?: string;
-  retentionPeriod?: number;
-  retentionType?: 'indefinitely' | 'days' | 'weeks' | 'months' | 'years';
-  activities: Activity[];
-  activityFlows: ActivityFlow[];
-  theme: Theme;
-};
-
 export type Event = {
   startTime: string;
   endTime: string;
@@ -112,6 +52,5 @@ export type Event = {
 
 export type AppletsSchema = {
   applets: BaseSchema<{ result: Applet[]; count: number } | null>;
-  applet: BaseSchema<{ result: SingleApplet } | null>;
   events: BaseSchema<{ result: Event[]; count: number } | null>;
 };
