+++ src/modules/Dashboard/state/Applets/Applets.state.ts
@@ -9,6 +9,5 @@ const initialStateData = {
 
 export const state: AppletsSchema = {
   applets: initialStateData,
-  applet: initialStateData,
   events: initialStateData,
 };
