+++ src/modules/Dashboard/state/index.ts
@@ -1,5 +1,4 @@
 export * from './Account';
-export * from './Applet';
 export * from './Applets';
 export * from './Folders';
 export * from './Popups';
