+++ src/shared/api/api.ts
@@ -1,5 +1,5 @@
 import { apiClient, authApiClient } from './api.client';
-import { SignInRefreshTokenArgs } from './api.types';
+import { SignInRefreshTokenArgs, AppletId } from './api.types';
 
 export const signInRefreshTokenApi = (
   { refreshToken }: SignInRefreshTokenArgs,
@@ -17,3 +17,6 @@ export const getWorkspacesApi = (signal?: AbortSignal) =>
   authApiClient.get('/workspaces', {
     signal,
   });
+
+export const getAppletApi = ({ appletId }: AppletId, signal?: AbortSignal) =>
+  authApiClient.get(`/applets/${appletId}`, { signal });
