+++ src/shared/state/Applet/Applet.reducer.ts
@@ -1,0 +1,16 @@
+import { ActionReducerMapBuilder } from '@reduxjs/toolkit';
+
+import { AppletSchema } from './Applet.schema';
+import { getApplet } from './Applet.thunk';
+
+import {
+  createAppletPendingData,
+  createAppletFulfilledData,
+  createAppletRejectedData,
+} from './Applet.utils';
+
+export const extraReducers = (builder: ActionReducerMapBuilder<AppletSchema>): void => {
+  createAppletPendingData({ builder, thunk: getApplet, key: 'applet' });
+  createAppletFulfilledData({ builder, thunk: getApplet, key: 'applet' });
+  createAppletRejectedData({ builder, thunk: getApplet, key: 'applet' });
+};
