+++ src/shared/state/index.ts
@@ -1,2 +1,3 @@
 export * from './Base';
 export * from './Breadcrumbs';
+export * from './Applet';
